Add project setup instructions to README.md and update ALLOWED_HOSTS in settings.py
New instructions for project setup were added in README.md, providing a simple guide on how to clone and start the project. This will help new contributors to quickly set up and run the project.

Additionally, settings.py was modified to update ALLOWED_HOSTS to accept any host, since the prior condition restricted connection to localhost only. This makes the server publicly accessible, enabling the testing and usage of the application on different platforms/networks.